fix: file delete

+++ src/infra/helpers/file-delete.ts
@@ -1,14 +1,11 @@
 import { unlink } from 'fs';
 
 const deleteFile = (dirname: string): void => {
-  unlink(dirname, (err) => {
-    try {
+  if (dirname) {
+    unlink(dirname, (err) => {
       if (err) console.log(err);
-      console.log(dirname, ' was deleted');
-    } catch (err) {
-      console.log(err);
-    }
-  });
+    });
+  }
 };
 
 export default deleteFile;
